migrate to flink 1.15 and scala 2.13 (#432)
* migrate to flink 1.15 and scala 2.13

* fix assembly conflicts

* switch e2e tests to 2.13

+++ src/main/scala/ai/metarank/feature/ItemAgeFeature.scala
@@ -67,7 +67,7 @@ case class ItemAgeFeature(schema: ItemAgeSchema) extends ItemFeature with Loggin
     features.get(Key(conf, Tenant(request.tenant), id.id.value)) match {
       case Some(ScalarValue(_, _, SDouble(value))) =>
         val updatedAt = Timestamp(math.round(value * 1000))
-        SingleValue(schema.name, updatedAt.diff(request.timestamp).toSeconds)
+        SingleValue(schema.name, updatedAt.diff(request.timestamp).toSeconds.toDouble)
       case _ => SingleValue(schema.name, 0.0)
     }
